[#4] 코드 수정
